return same result when called twice
There is a bug in Nokogiri
(https://github.com/sparklemotion/nokogiri/issues/1282) that causes
`Nokogiri::XML::RelaxNG#validate` in JRuby to add errors every
time it's called.

To fix this, create a new instance of `Nokogiri::XML::RelaxNG` every
time. To lessen the impact on the library's performance, use
`from_document` and cache the parsed XML.